Stage 5.1
1. Create a Node.js server app and move the following logic
from the frontend app from the previous stage:
a. POST request for login (auth is still a ‘simulation’)
b. GET request to get/search a list of projects

2. Add Prettier to your project as a pre-commit hook

3. Add a unit testing library to your project and cover application with the following test
scenarios:
a. Mock request to the server (in unit tests, we shouldn’t
depend on a server, so we simulate its response via
jest.mock()).
b. Write a first unit test that covers the scenario when auth
is a success, then success action is dispatched,
c. Write a second unit test that covers the scenario when
auth is failed, then error/fail action is dispatch

4. Add a pre-push hook to execute tests before git push. If tests
have failed, git push shouldn’t be executed

5. Set up Node.js debugging

6. Set up Jest debugging